Finished main menu

+++ Assets/qASIC/Options/OptionsSlider.cs
@@ -31,7 +31,7 @@ namespace qASIC.Options.Menu
                 SetValue(intResult, log);
                 return;
             }
-            SetValue(round ? Mathf.Round(value / (float)roundValue) * roundValue : value, log);
+            SetValue(round ? (float)(Mathf.Round(value / (float)roundValue) * roundValue) : value, log);
         }
 
         public override string GetLabel()
